authorization demo added

+++ api/app/Http/Controllers/ProjectsController.php
@@ -31,24 +31,37 @@ class ProjectsController extends Controller
 
     public function show(Project $project)
     {
+        // abort_if($project->owner_id !== auth()->id(), 403); // Laravel help function
+        // abort_if(! auth()->user()->owns($project), 403);     // Object oriented, add own method to user class
+        // abort_unless( auth()->user()->owns($project), 403);
+
+        
+        // if (\Gate::denies('update', $project)) // or Gate::allows
+        // {
+            //     abort(403);
+            // }
+            
+        $this->authorize('update', $project); // With ProjectPolicy.php, registered on AuthServiceProvider.php
+
         return view('projects.show', compact('project'));
     }
 
     public function store ()
     {
-        // $attributes = request()->validate([
-        //     'title' => ['required', 'min:3'],
-        //     'description' => ['required', 'min:3']
-        // ]);
+        $attributes = request()->validate([
+            'title' => ['required', 'min:3'],
+            'description' => ['required', 'min:3']
+        ]);
 
-        // Project::create($attributes);
+        $attributes ['owner_id'] = auth()->id();
+        Project::create($attributes);
 
-        Project::create (
-            request()->validate([
-                'title' => ['required', 'min:3'],
-                'description' => ['required', 'min:3']
-            ])
-        );
+        // Project::create (
+        //     request()->validate([
+        //         'title' => ['required', 'min:3'],
+        //         'description' => ['required', 'min:3']
+        //     ])
+        // );
 
         return redirect('/projects');
     }
@@ -60,6 +73,7 @@ class ProjectsController extends Controller
 
     public function update(Project $project)
     {
+        $this->authorize('update', $projects);
 
         $project->update(request(['title', 'description']));
 
@@ -68,6 +82,8 @@ class ProjectsController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('update', $projects);
+        
         $project->delete();
 
         return redirect('/projects');
